requirements.txt
aiogram
psycopg2
environs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client